refactor: Split etcRouter/Controller into chart and lastExpense
